Refactor some constants and functions to improve reusability
