+++ internal/common/root_install.go
@@ -1,4 +1,4 @@
-package installresources
+package common
 
 import (
 	"context"
@@ -41,7 +41,7 @@ func (i *RootInstall) Create(ctx context.Context, diags *diag.Diagnostics, plan 
 
 	item := i.FromJson(ctx, diags, json)
 	if item == nil {
-		reportConversionError("Bad API response", "Could not read resource data from", json, diags)
+		ReportConversionError("Bad API response", "Could not read resource data from", json, diags)
 		return
 	}
 
@@ -67,7 +67,7 @@ func (i *RootInstall) Read(ctx context.Context, diags *diag.Diagnostics, state *
 
 	item := i.FromJson(ctx, diags, json)
 	if item == nil {
-		reportConversionError("Bad API response", "Could not read resource data from", json, diags)
+		ReportConversionError("Bad API response", "Could not read resource data from", json, diags)
 		return
 	}
