+++ internal/provider/resources/install/gcp/iam_write.go
@@ -7,6 +7,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/p0-security/terraform-provider-p0/internal"
+	"github.com/p0-security/terraform-provider-p0/internal/common"
 	installresources "github.com/p0-security/terraform-provider-p0/internal/provider/resources/install"
 )
 
@@ -20,7 +21,7 @@ func NewGcpIamWrite() resource.Resource {
 }
 
 type GcpIamWrite struct {
-	installer *installresources.Install
+	installer *common.Install
 }
 
 func (r *GcpIamWrite) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
