+++ internal/provider/resources/install/resources.go
@@ -1,0 +1,6 @@
+package installresources
+
+const (
+	IamAssessment = "iam-assessment"
+	IamWrite      = "iam-write"
+)
